Restructure codebase to remove /utils
The distinction between /lib and /lib/utils modules is kind of arbitrary and unnecessary.
