+++ lib/commands/remove.ts
@@ -3,7 +3,7 @@ import { ParsedMail } from "mailparser";
 import config from "../config";
 import { email, operationalDomain } from "../env";
 import { Commands } from "../commandSet";
-import sendEmail from "../utils/sendEmail";
+import sendEmail from "../sendEmail";
 
 export default async (parsedMail: ParsedMail): Promise<void> => {
   const providedAlias = parsedMail.subject;
