+++ tests/lib/commands/list.test.ts
@@ -2,11 +2,11 @@ import path from "path";
 import * as AWSMock from "aws-sdk-mock";
 import { DynamoDB } from "aws-sdk";
 import list from "../../../lib/commands/list";
-import sendEmail from "../../../lib/utils/sendEmail";
+import sendEmail from "../../../lib/sendEmail";
 import { email, operationalDomain } from "../../../lib/env";
 import { Commands } from "../../../lib/commandSet";
 
-jest.mock("../../../lib/utils/sendEmail");
+jest.mock("../../../lib/sendEmail");
 
 type Callback = (err: any, data: any) => void;
